Pipeline_Files/requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# API and encryption
requests>=2.28.0
pycryptodome>=3.15.0

# Data processing
joblib>=1.1.0

# Airflow (usually pre-installed in Airflow environment)
apache-airflow>=2.5.0

# Data visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.10.0

# Advanced ML & Deep Learning
tensorflow>=2.10.0  # For deep learning
xgboost>=1.6.0     # For gradient boosting
lightgbm>=3.3.0    # For gradient boosting

# Data handling & processing
scipy>=1.9.0       # Scientific computing
statsmodels>=0.13.0  # Statistical models
python-dotenv>=0.21.0  # Environment variable management

# Optional: for better performance
openpyxl>=3.0.10  # For Excel file support
fastparquet>=0.8.0  # For parquet file support
python-snappy>=0.6.0  # For compression support